First steps in merge cluster list function - at the moment, loops over and finds free clusters and pulls them out of these lists. Creates a separate 'to merge' list.
PROBLEM! Depth first search is not calculating the bounds properly -> I think this is because there is no wrapping, therefore cannot actually check if it is connected to the next segment/ can add to bound. Need to do 2 separate 'if' statements - firstly to see if connected to the upper WRAPPED segment (and if so, add to the bounds), then next, to check if it is inside the bounds, and if so, to continue the depth first search.