BUGFIX: MPI Process was hitting an error because of c memcpy function, NOT because of DFS. Replacing this with a loop, and removing all MPI_Barrier()s in the DFS (these will break it), allows it to run smoothly.
NEXT: Need to put cluster merging back in (should work easily as no different to MPI). Then each MPI node will have a linked list of clusters. Need to be able to send these to the master process - will probably have to keep track of the number of nodes in the linked list in order to convert it into an array. May need to start freeing data or will run out.